requirements.txt
faiss-cpu
cohere
sentence-transformers
pdfplumber
gradio